+++ v2_user_config_validator.go
@@ -3,6 +3,7 @@ package userconfig
 import (
 	"encoding/json"
 	"fmt"
+	"os"
 	"sort"
 	"strconv"
 	"strings"
@@ -429,4 +430,34 @@ func (nds ComponentDefinitions) validateLeafs() error {
 		}
 	}
 	return nil
+}
+
+// getArrayEntry tries to get an entry in the given map that is an array of
+// objects.
+func getArrayEntry(def map[string]interface{}, key string) []interface{} {
+	entry, ok := def[key]
+	if !ok {
+		// No key element found
+		return nil
+	}
+
+	entryArr, ok := entry.([]interface{})
+	if !ok {
+		// entry not right type
+		return nil
+	}
+
+	return entryArr
+}
+
+// normalizeFolder removes any trailing path separator from the given path.
+func normalizeFolder(path string) string {
+	if path == "" {
+		return ""
+	}
+	l := len(path)
+	if os.IsPathSeparator(path[l-1]) {
+		path = path[:l-1]
+	}
+	return path
 }
